refactor(reducers): move `q`
