+++ app/renderer/src/reducers/q/index.js
@@ -1,0 +1,1 @@
+export { default } from './q';
